# src/apiserver/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V2 API",
    "version": "2.0.0",
    "description": "Enterprise platform API for PulseOps V2. Supports database management, authentication (JWT + HttpOnly cookies), module hot-dropping, and system configuration. All URLs are defined in urls.json — this spec references relative paths under the /api prefix.",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative to host — works on any environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via POST /auth/login. Expiry: 24h."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie set automatically on login. Same JWT as BearerAuth."
      }
    }
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new access token. Also sets updated HttpOnly cookies. Use this when the access token expires (24h) but the refresh token is still valid (7d).",
        "requestBody": {
          "required": true,
          "description": "The refresh token received from the login response.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" } } }
        },
        "responses": {
          "200": {
            "description": "New access token generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": { "type": "string" },
                        "expiresIn": { "type": "integer", "example": 86400 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Refresh token is invalid or expired.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/system/config/{key}": {
      "get": {
        "tags": ["Configuration"],
        "operationId": "getConfig",
        "summary": "Get config by key",
        "description": "Returns a single configuration entry by its key. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "key", "in": "path", "required": true, "description": "Config key to retrieve", "schema": { "type": "string" }, "example": "auth_provider" }
        ],
        "responses": {
          "200": { "description": "Config entry.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Key not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Configuration"],
        "operationId": "deleteConfig",
        "summary": "Delete config by key",
        "description": "Permanently deletes a configuration entry. Requires super_admin role.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "key", "in": "path", "required": true, "description": "Config key to delete", "schema": { "type": "string" }, "example": "auth_provider" }
        ],
        "responses": {
          "200": { "description": "Config deleted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Requires super_admin role.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Key not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  }
}